+++ src/main/java/ru/ai/sin/service/impl/SkillService.java
@@ -11,7 +11,7 @@ public interface SkillService {
     // ---------- GET METHODS ----------
     SkillDTO getById(long id);
 
-    List<SkillDTO> getAll(long page, long size);
+    List<SkillDTO> getAll(int page, int size);
 
     // ---------- POST METHODS ----------
     SkillDTO create(AddSkillReq addSkillReq);
